requirements.txt
fastapi
pydantic
uvicorn
sqlalchemy
alembic
PyJWT
python-dotenv
elasticsearch
pytest
httpx
httptools
uvloop
fastapi_pagination
psycopg

